requirements.txt
apache-airflow==2.10.0
numpy==2.1.0
requests==2.32.3